dashboard/backend/requirements.txt
fastapi
uvicorn[standard]
databricks-sql-connector
pydantic
pydantic-settings
python-dotenv
